+++ include/alphablend.h
@@ -1,6 +1,6 @@
 #ifndef _ALPHABLEND_H
 #define _ALPHABLEND_H
-#define WINVER 0x0500
+
 #include <windows.h>
 
 BOOL c_AlphaBlend ( HDC hdcDest, int nXOriginDest, int nYOriginDest, int nWidthDest, int hHeightDest